requirements.txt
python-dotenv>=1.0.0
google-generativeai>=0.3.0

# LangGraph for orchestration
langgraph>=0.0.26
langchain-core>=0.1.0

# Code analysis tools
pylint>=3.0.0
astroid>=3.0.0

# Testing framework
pytest>=7.4.0

# Utility libraries
pydantic>=2.5.0
typing-extensions>=4.9.0